requirements.txt
streamlit
python-dotenv==1.0.0
PyPDF2
langchain==0.0.284
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=3.19.0
InstructorEmbedding
torch
sentence_transformers==2.2.2
google-generativeai